* Updated code-gen relative paths for the   refactor.
